test(webui): exercise VM harness production seam
