+++ crates/product/ironclaw_webui/frontend/src/pages/settings/vm-module-harness.test.ts
@@ -1,53 +1,34 @@
 // @ts-nocheck
 import assert from "node:assert/strict";
 import { test } from "vitest";
-import vm from "node:vm";
 
-import { sourceTextForVmTest } from "../../test-support/vm-module-harness";
+import {
+  runVmModuleForTest,
+  sourceForVmTest,
+} from "../../test-support/vm-module-harness";
 
-function evaluate(source, exportNames) {
-  const context = { globalThis: {} };
-  const script = new vm.Script(sourceTextForVmTest(source, exportNames));
-  script.runInNewContext(context);
-  return context.globalThis.__testExports;
+const FIXTURE_PATH = "./vm-module-harness.fixture.ts";
+
+function runFixture(exportNames) {
+  return runVmModuleForTest(FIXTURE_PATH, exportNames, {}, import.meta.url);
 }
 
 test("VM harness keeps code after semicolonless multiline imports", () => {
-  const exports = evaluate(
-    `
-import {
-  alpha,
-  beta,
-} from "./deps.js"
-const value = "still here";
-export function readValue() {
-  return value;
-}
-`,
-    ["readValue"]
-  );
+  const exports = runFixture(["readValue"]);
 
   assert.equal(exports.readValue(), "still here");
 });
 
 test("VM harness captures exported declarations and named export aliases", () => {
-  const source = `
-const base = 41;
-export interface AnswerMetadata {
-  source: string;
-}
-export type Answer = number;
-export const answer = base + 1;
-function readAnswer() {
-  return answer;
-}
-export { readAnswer as getAnswer };
-`;
   const exportNames = ["answer", "getAnswer"];
-  const executableSource = sourceTextForVmTest(source, exportNames);
+  const executableSource = sourceForVmTest(
+    FIXTURE_PATH,
+    exportNames,
+    import.meta.url,
+  );
 
   assert.doesNotMatch(executableSource, /^\s*(?:interface|type)\b/m);
-  const exports = evaluate(source, exportNames);
+  const exports = runFixture(exportNames);
 
   assert.equal(exports.answer, 42);
   assert.equal(exports.getAnswer(), 42);
